requirements.txt
asgiref==3.8.1
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
decorator==4.4.2
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
moviepy==2.1.1
msgpack==1.1.0
numpy==2.2.1
pillow==10.4.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
pytube==15.0.0
pytubefix==8.10.2
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
tqdm==4.67.1
uritemplate==4.1.1
urllib3==2.3.0
